[autothrottle] comments unused code
